CD-207: Revise statistics retrieval script
- Separate functions and group them into classes based on their related functionality
- Adds functions for retrieving Single-cell-related statistics
- Improve readability and making it easier to add additional statistical functionality by class
- Rename the script from `get_sample_statistics.py` to `get_statistics.py` because its contents does not only pertain to samples
